Enhance cross-version support and migrate to Scala 3
- Implement cross-version testing using `crossTest` projects against published artifacts.
- Introduce code refactoring to replace non-local returns with `scala.util.boundary` for Scala 3 compatibility.
- Improve type safety with float-to-double conversions and refine parameter extraction logic.
- Add `commonSettings` for streamlined cross-compilation and update test structures for better robustness.
- Update documentation with guidelines for cross-compilation and testing.

+++ src/main/scala/org/llm4s/llmconnect/provider/OpenAIClient.scala
@@ -30,11 +30,11 @@ class OpenAIClient(config: OpenAIConfig) extends LLMClient {
       val chatOptions = new ChatCompletionsOptions(chatMessages)
 
       // Set options
-      chatOptions.setTemperature(options.temperature.floatValue())
+      chatOptions.setTemperature(options.temperature.doubleValue())
       options.maxTokens.foreach(mt => chatOptions.setMaxTokens(mt))
-      chatOptions.setPresencePenalty(options.presencePenalty.floatValue())
-      chatOptions.setFrequencyPenalty(options.frequencyPenalty.floatValue())
-      chatOptions.setTopP(options.topP.floatValue())
+      chatOptions.setPresencePenalty(options.presencePenalty.doubleValue())
+      chatOptions.setFrequencyPenalty(options.frequencyPenalty.doubleValue())
+      chatOptions.setTopP(options.topP.doubleValue())
 
       // Add tools if specified
       if (options.tools.nonEmpty) {
